Fixing int to uint in user commands

+++ Application/Users/Commands/DeleteUser.cs
@@ -4,13 +4,13 @@ using MediatR;
 
 namespace Application.Users.Commands;
 
-public record DeleteUserCommand(int Id) : IRequest;
+public record DeleteUserCommand(uint Id) : IRequest;
 
 public class DeleteUserCommandValidator : AbstractValidator<DeleteUserCommand>
 {
     public DeleteUserCommandValidator()
     {
-        RuleFor(x => x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
+        RuleFor(x => (int)x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
     }
 }
 
@@ -18,6 +18,6 @@ public class DeleteUserHandler(IUserService userService) : IRequestHandler<Delet
 {
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
-        await userService.DeleteUserAsync((uint)request.Id);
+        await userService.DeleteUserAsync(request.Id);
     }
 }
