+++ Application/Users/Commands/UpdateUser.cs
@@ -5,13 +5,13 @@ using MediatR;
 
 namespace Application.Users.Commands;
 
-public record UpdateUserCommand(int Id, UpdateUserDto User) : IRequest;
+public record UpdateUserCommand(uint Id, UpdateUserDto User) : IRequest;
 
 public class UpdateUserCommandValidator : AbstractValidator<UpdateUserCommand>
 {
     public UpdateUserCommandValidator()
     {
-        RuleFor(x => x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
+        RuleFor(x => (int)x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
         RuleFor(x => x.User.Phone)
             .NotEmpty().WithMessage("Phone number is required.")
             .Matches(@"^\+?[1-9]\d{1,14}$").WithMessage("Phone must be in a valid international format.");
@@ -22,6 +22,6 @@ public class UpdateUserHandler(IUserService userService) : IRequestHandler<Updat
 {
     public async Task Handle(UpdateUserCommand request, CancellationToken cancellationToken)
     {
-        await userService.UpdateUserAsync((uint)request.Id, request.User);
+        await userService.UpdateUserAsync(request.Id, request.User);
     }
 }
